requirements.txt
asgiref==3.5.2
boto3==1.26.54
botocore==1.29.54
Django==4.1.1
django-admin==2.0.1
django-ckeditor==6.5.1
django-dotenv==1.4.2
django-environ==0.9.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==22.1
django-js-asset==2.0.0
django-six==1.0.4
django-storages==1.13.1
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.2.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.2.1
s3transfer==0.6.0
screen==1.0.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
whitenoise==6.2.0
xlwt==1.3.0
